Extensive refactoring continued.
