--- scripts/utils/listeners.js
@@ -1,205 +1,0 @@
-import { EquipmentParser, HM, SavedOptions, StatRoller, SummaryManager } from './index.js';
-
-/**
- * Manages event listeners and UI updates for the HeroMancer application.
- * Handles ability scores, equipment selection, character details, and UI summaries.
- * @class
- */
-export class Listeners {
-  /* -------------------------------------------- */
-  /*  Static Public Methods                       */
-  /* -------------------------------------------- */
-
-  /**
-   * Restores previously saved form options
-   * @param {HTMLElement} html - The form element
-   * @returns {Promise<void>}
-   * @static
-   */
-  static async restoreFormOptions(html) {
-    const savedOptions = await SavedOptions.loadOptions();
-    if (Object.keys(savedOptions).length === 0) return;
-
-    // First pass to restore all form elements
-    for (const [key, value] of Object.entries(savedOptions)) {
-      const elem = html.querySelector(`[name="${key}"]`);
-      if (!elem) continue;
-
-      if (elem.type === 'checkbox') {
-        elem.checked = value;
-      } else if (elem.tagName === 'SELECT') {
-        elem.value = value;
-
-        // Update HM.SELECTED for class, race, background
-        if (key === 'class' || key === 'race' || key === 'background') {
-          HM.SELECTED[key] = {
-            value: value,
-            id: value.split(' ')[0],
-            uuid: value.match(/\[(.*?)]/)?.[1]
-          };
-        }
-      } else {
-        elem.value = value;
-      }
-    }
-
-    // Force updates after restoring options
-    requestAnimationFrame(() => {
-      SummaryManager.updateClassRaceSummary();
-
-      // Update equipment if needed
-      if (!HM.COMPAT.ELKAN) {
-        const equipmentContainer = html.querySelector('#equipment-container');
-        if (equipmentContainer) {
-          const equipment = new EquipmentParser();
-          equipment
-            .generateEquipmentSelectionUI()
-            .then((choices) => {
-              equipmentContainer.innerHTML = '';
-              equipmentContainer.appendChild(choices);
-              DOMManager.attachEquipmentListeners(equipmentContainer);
-            })
-            .catch((error) => HM.log(1, 'Error rendering equipment choices:', error));
-        }
-      }
-    });
-  }
-
-  /**
-   * Updates the display of remaining points in the abilities tab
-   * @param {number} remainingPoints - The number of points remaining to spend
-   * @static
-   */
-  static updateRemainingPointsDisplay(remainingPoints) {
-    const abilitiesTab = document.querySelector(".tab[data-tab='abilities']");
-    if (!abilitiesTab?.classList.contains('active')) return;
-
-    const remainingPointsElement = document.getElementById('remaining-points');
-    const totalPoints = StatRoller.getTotalPoints();
-
-    if (remainingPointsElement) {
-      remainingPointsElement.innerHTML = remainingPoints;
-      this.#updatePointsColor(remainingPointsElement, remainingPoints, totalPoints);
-    }
-  }
-
-  /**
-   * Adjusts ability score up or down within valid range and point limits
-   * @param {number} index - The index of the ability score to adjust
-   * @param {number} change - The amount to change the score by (positive or negative)
-   * @param {number[]} selectedAbilities - Array of current ability scores
-   * @static
-   */
-  static changeAbilityScoreValue(index, change, selectedAbilities) {
-    if (!Array.isArray(selectedAbilities)) {
-      HM.log(1, 'selectedAbilities must be an array');
-      return;
-    }
-    const abilityScoreElement = document.getElementById(`ability-score-${index}`);
-    const currentScore = parseInt(abilityScoreElement.innerHTML, 10);
-    const { MIN, MAX } = HM.ABILITY_SCORES;
-    const newScore = Math.min(MAX, Math.max(MIN, currentScore + change));
-    const totalPoints = StatRoller.getTotalPoints();
-    const pointsSpent = StatRoller.calculateTotalPointsSpent(selectedAbilities);
-
-    if (change > 0 && pointsSpent + StatRoller.getPointBuyCostForScore(newScore) - StatRoller.getPointBuyCostForScore(currentScore) > totalPoints) {
-      HM.log(2, 'Not enough points remaining to increase this score.');
-      return;
-    }
-
-    if (newScore !== currentScore) {
-      abilityScoreElement.innerHTML = newScore;
-      selectedAbilities[index] = newScore;
-
-      const updatedPointsSpent = StatRoller.calculateTotalPointsSpent(selectedAbilities);
-      const remainingPoints = totalPoints - updatedPointsSpent;
-
-      this.updateRemainingPointsDisplay(remainingPoints);
-      this.updatePlusButtonState(selectedAbilities, remainingPoints);
-      this.updateMinusButtonState(selectedAbilities);
-    }
-  }
-
-  /**
-   * Updates the state of plus buttons based on available points and maximum scores
-   * @param {number[]} selectedAbilities - Array of current ability scores
-   * @param {number} remainingPoints - Points available to spend
-   * @static
-   */
-  static updatePlusButtonState(selectedAbilities, remainingPoints) {
-    // Create a document fragment for batch processing
-    const updates = [];
-    const { MAX } = HM.ABILITY_SCORES;
-
-    document.querySelectorAll('.plus-button').forEach((button, index) => {
-      const currentScore = selectedAbilities[index];
-      const pointCostForNextIncrease = StatRoller.getPointBuyCostForScore(currentScore + 1) - StatRoller.getPointBuyCostForScore(currentScore);
-      const shouldDisable = currentScore >= MAX || remainingPoints < pointCostForNextIncrease;
-
-      // Only update if the state actually changes
-      if (button.disabled !== shouldDisable) {
-        updates.push(() => (button.disabled = shouldDisable));
-      }
-
-      const inputElement = document.getElementById(`ability-${index}-input`);
-      if (inputElement && inputElement.value !== String(currentScore)) {
-        updates.push(() => (inputElement.value = currentScore));
-      }
-    });
-
-    // Apply all updates in one batch
-    if (updates.length) {
-      requestAnimationFrame(() => updates.forEach((update) => update()));
-    }
-  }
-
-  /**
-   * Updates the state of minus buttons based on minimum allowed scores
-   * @param {number[]} selectedAbilities - Array of current ability scores
-   * @static
-   */
-  static updateMinusButtonState(selectedAbilities) {
-    const updates = [];
-    const { MIN } = HM.ABILITY_SCORES;
-
-    document.querySelectorAll('.minus-button').forEach((button, index) => {
-      const currentScore = selectedAbilities[index];
-      const shouldDisable = currentScore <= MIN;
-
-      // Only update if the state actually changes
-      if (button.disabled !== shouldDisable) {
-        updates.push(() => (button.disabled = shouldDisable));
-      }
-
-      const inputElement = document.getElementById(`ability-${index}-input`);
-      if (inputElement && inputElement.value !== String(currentScore)) {
-        updates.push(() => (inputElement.value = currentScore));
-      }
-    });
-
-    // Apply all updates in one batch
-    if (updates.length) {
-      requestAnimationFrame(() => updates.forEach((update) => update()));
-    }
-  }
-
-  /* -------------------------------------------- */
-  /*  Static Private Methods                      */
-  /* -------------------------------------------- */
-
-  /**
-   * Updates the color of the remaining points display based on percentage remaining
-   * @param {HTMLElement} element - The element to update
-   * @param {number} remainingPoints - Current remaining points
-   * @param {number} totalPoints - Total available points
-   * @private
-   * @static
-   */
-  static #updatePointsColor(element, remainingPoints, totalPoints) {
-    if (!element) return;
-
-    const percentage = (remainingPoints / totalPoints) * 100;
-    const hue = Math.max(0, Math.min(120, (percentage * 120) / 100));
-    element.style.color = `hsl(${hue}, 100%, 35%)`;
-  }
-}
